+++ Source/Alternet.UI/Controls/BaseControls/Control.Int.Methods.cs
@@ -42,7 +42,7 @@ namespace Alternet.UI
 
         internal static SizeD GetPreferredSizeDefaultLayout(Control container, SizeD availableSize)
         {
-            if (container.HasChildren || container.Handler.HasVisualChildren)
+            if (container.HasChildren /*|| container.Handler.HasVisualChildren*/)
                 return container.Handler.GetSpecifiedOrChildrenPreferredSize(availableSize);
             return container.Handler.GetNativeControlSize(availableSize);
         }
@@ -200,6 +200,12 @@ namespace Alternet.UI
             {
                 OnMouseRightButtonDown(e);
             }
+        }
+
+        internal void RaiseProcessException(ControlExceptionEventArgs e)
+        {
+            OnProcessException(e);
+            ProcessException?.Invoke(this, e);
         }
 
         internal void RaiseMouseWheel(MouseEventArgs e)