+++ Source/Alternet.UI/Controls/BaseControls/Control.Pro.Methods.cs
@@ -63,6 +63,15 @@ namespace Alternet.UI
         }
 
         /// <summary>
+        /// Raises the <see cref="Idle"/> event.
+        /// </summary>
+        /// <param name="e">An <see cref="EventArgs"/> that contains the event
+        /// data.</param>
+        protected virtual void OnIdle(EventArgs e)
+        {
+        }
+
+        /// <summary>
         /// Raises the <see cref="Control.Resize" /> event.
         /// </summary>
         /// <param name="e">An <see cref="EventArgs" /> that contains the event data.</param>
@@ -91,8 +100,7 @@ namespace Alternet.UI
 
         /// <summary>
         /// Called when a <see cref="Control"/> is inserted into
-        /// the <see cref="Control.Children"/> or
-        /// <see cref="ControlHandler.VisualChildren"/> collection.
+        /// the <see cref="Control.Children"/>.
         /// </summary>
         protected virtual void OnChildInserted(Control childControl)
         {
@@ -170,15 +178,6 @@ namespace Alternet.UI
         /// <param name="e">An <see cref="EventArgs"/> that contains the event
         /// data.</param>
         protected virtual void OnClick(EventArgs e)
-        {
-        }
-
-        /// <summary>
-        /// Called when the application is in idle state.
-        /// </summary>
-        /// <param name="e">An <see cref="EventArgs"/> that contains the event
-        /// data.</param>
-        protected virtual void OnIdle(EventArgs e)
         {
         }
 
@@ -305,8 +304,8 @@ namespace Alternet.UI
             SuspendLayout();
             if (HasChildren)
                 Children.Clear();
-            if (Handler.HasVisualChildren)
-                Handler.VisualChildren.Clear();
+            /*if (Handler.HasVisualChildren)
+                Handler.VisualChildren.Clear();*/
             ResumeLayout(performLayout: false);
 
             // TODO
@@ -468,6 +467,15 @@ namespace Alternet.UI
         }
 
         /// <summary>
+        /// Called when an exception need to be processed.
+        /// </summary>
+        /// <param name="e">An <see cref="ControlExceptionEventArgs"/> that contains
+        /// the event data.</param>
+        protected virtual void OnProcessException(ControlExceptionEventArgs e)
+        {
+        }
+
+        /// <summary>
         ///     Virtual method reporting a mouse move
         /// </summary>
         protected virtual void OnMouseMove(MouseEventArgs e)
@@ -498,7 +506,6 @@ namespace Alternet.UI
         /// <param name="e">Event arguments.</param>
         protected virtual void OnToolTipChanged(EventArgs e)
         {
-            ToolTipChanged?.Invoke(this, e);
         }
 
         /// <summary>