+++ Source/Samples/ControlsSample/Misc/MainWindow.cs
@@ -20,6 +20,8 @@ namespace ControlsSample
 
         public MainWindow()
         {
+            SupressEsc = true;
+
             eventsControl = new()
             {
                 Dock = DockStyle.Bottom,
@@ -128,10 +130,7 @@ namespace ControlsSample
         {
             NameValue<Func<Control>>? popupNameValue;
 
-            if (!Application.IsWindowsOS)
-                popupNameValue = null;
-            else
-                popupNameValue = new("Popup", () => new ListControlsPopups());
+            popupNameValue = new("Popup", () => new ListControlsPopups());
 
             NameValue<Func<Control>>?[] pages =
             {